fix(agent-tars-cli): init api
