+++ multimodal/agent-tars-cli/src/cli.ts
@@ -4,22 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import cac from 'cac';
-import { printWelcomeLogo } from './utils';
-import { registerCommands } from './commands';
+import { bootstrapCli } from './index';
 
-// Display ASCII art LOGO immediately at program entry
-printWelcomeLogo();
+export { bootstrapCli };
 
-// Create CLI with custom styling
-const cli = cac('tars');
-
-// Use package.json version
-cli.version(__VERSION__);
-cli.help();
-
-// Register all commands
-registerCommands(cli);
-
-// Parse command line arguments
-cli.parse();
+if (require.main === module) {
+  bootstrapCli();
+}
