+++ multimodal/agent-tars-cli/src/index.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2025 Bytedance, Inc. and its affiliates.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import cac from 'cac';
+import { printWelcomeLogo } from './utils';
+import { registerCommands } from './commands';
+
+export function bootstrapCli() {
+  // Display ASCII art LOGO immediately at program entry
+  printWelcomeLogo();
+
+  // Create CLI with custom styling
+  const cli = cac('tars');
+
+  // Use package.json version
+  cli.version(__VERSION__);
+  cli.help();
+
+  // Register all commands
+  registerCommands(cli);
+
+  // Parse command line arguments
+  cli.parse();
+}
